requirements.txt
torch>=2.5.1
bitsandbytes>=0.45.0
faiss-gpu-cu12>=1.9.0.post1
llama-index>=0.12.5
llama-index-embeddings-huggingface>=0.4.0
llama-index-llms-huggingface>=0.4.1
llama-index-vector-stores-faiss>=0.3.0
pymupdf>=1.25.1
pytest>=8.3.4
streamlit>=1.41.1
